+++ packages/amplify_auth_cognito/example/ios/unit_tests/amplify_auth_cognito_tests.swift
@@ -876,12 +876,17 @@ class amplify_auth_cognito_tests: XCTestCase {
     
     func test_confirmSignInValidationOptions() {
         let rawData: NSMutableDictionary = ["confirmationCode": _confirmationCode]
-        let rawOptions: Dictionary<String, Any> = ["clientMetadata" : ["foo": "bar"]]
+        let rawOptions: Dictionary<String, Any> = [
+            "clientMetadata" : ["foo": "bar"],
+            "userAttributes": ["email": "test@test.test"]
+        ]
         rawData["options"] = rawOptions
         XCTAssertNoThrow(try FlutterConfirmSignInRequest.validate(dict: rawData))
         let req = FlutterConfirmSignInRequest(dict: rawData)
         let options = (req.options?.pluginOptions as! AWSAuthConfirmSignInOptions)
         XCTAssertEqual(options.metadata, ["foo": "bar"])
+        XCTAssertEqual(options.userAttributes?[0].key, .email)
+        XCTAssertEqual(options.userAttributes?[0].value, "test@test.test")
     }
     
     func test_confirmSignInValidationNoOptions() {
@@ -1404,7 +1409,7 @@ class amplify_auth_cognito_tests: XCTestCase {
                 XCTAssertEqual(false, res.toJSON()["isSignedIn"] as? Bool)
                 XCTAssertEqual("testid", res.toJSON()["identityId"] as? String)
                 // userSub error will result in map with one 'error' key
-                XCTAssertEqual(1, (res.toJSON()["tokens"] as?  [String: String])!.count)
+                XCTAssertNil(res.toJSON()["tokens"] as?  [String: String])
                 // credentials map should have access key and secret key
                 XCTAssertEqual(2, (res.toJSON()["credentials"] as?  [String: String])!.count)
             } else {
@@ -1449,7 +1454,7 @@ class amplify_auth_cognito_tests: XCTestCase {
                 // all tokens should be present with userpool-only access
                 XCTAssertEqual(3, (res.toJSON()["tokens"] as?  [String: String])!.count)
                 // credentials map should be empty
-                XCTAssertEqual(0, (res.toJSON()["credentials"] as?  [String: String])!.count)
+                XCTAssertNil(res.toJSON()["credentials"] as?  [String: String])
             } else {
                 XCTFail()
             }