+++ packages/amplify_auth_plugin_interface/lib/src/Session/AuthSession.dart
@@ -13,13 +13,9 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_auth_plugin_interface/amplify_auth_plugin_interface.dart';
-import 'package:flutter/foundation.dart';
-
-
 class AuthSession {
-  bool isSignedIn;
-  AuthSession({@required this.isSignedIn}) {
+  late bool isSignedIn;
+  AuthSession({required this.isSignedIn}) {
     this.isSignedIn = isSignedIn;
-   }
+  }
 }