+++ packages/amplify_auth_plugin_interface/lib/src/SignUp/SignUpRequest.dart
@@ -14,21 +14,22 @@
  */
 
 import './SignUpOptions.dart';
-import 'package:flutter/foundation.dart';
 
 class SignUpRequest {
   String username;
   String password;
-  SignUpOptions options;
+  SignUpOptions? options;
 
-  SignUpRequest({@required this.username, @required this.password, this.options});
+  SignUpRequest({required this.username, required this.password, this.options});
   Map<String, dynamic> serializeAsMap() {
-    final Map<String, dynamic> pendingRequest = <String, dynamic>{};
-    pendingRequest['username'] = username;
-    pendingRequest['password'] = password;
+    final Map<String, dynamic> pendingRequest = {
+      'username': username,
+      'password': password,
+    };
     if (options != null) {
-      pendingRequest['options'] = options.serializeAsMap();
+      pendingRequest['options'] = options!.serializeAsMap();
     }
+    pendingRequest.removeWhere((_, v) => v == null);
     return pendingRequest;
   }
 }