+++ packages/amplify_auth_plugin_interface/lib/src/SignUp/SignUpResult.dart
@@ -13,13 +13,13 @@
  * permissions and limitations under the License.
  */
 
+
 import './AuthNextSignUpStep.dart';
-import 'package:flutter/foundation.dart';
 
 class SignUpResult { 
   bool isSignUpComplete;
   AuthNextSignUpStep nextStep;
-  SignUpResult({@required this.isSignUpComplete, @required this.nextStep}) {
+  SignUpResult({required this.isSignUpComplete, required this.nextStep}) {
     this.isSignUpComplete = isSignUpComplete;
     this.nextStep = nextStep;
   }