+++ packages/amplify_datastore/lib/types/DataStoreHubEvents/OutboxMutationEvent.dart
@@ -13,17 +13,19 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_datastore/types/DataStoreHubEvents/HubEventElementWithMetadata.dart';
 import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:amplify_core/types/hub/HubEventPayload.dart';
 import 'HubEventElement.dart';
 
 class OutboxMutationEvent extends HubEventPayload {
-  HubEventElement element;
-  String modelName;
+  late HubEventElement element;
+  late String modelName;
 
-  OutboxMutationEvent(Map<dynamic, dynamic> serializedData, ModelProviderInterface provider, String eventName) {
-    element = eventName == "outboxMutationEnqueued" ? HubEventElement(serializedData, provider) : HubEventElementWithMetadata(serializedData, provider) ;
+  OutboxMutationEvent(Map<dynamic, dynamic> serializedData,
+      ModelProviderInterface provider, String eventName) {
+    element = eventName == "outboxMutationEnqueued"
+        ? HubEventElement.fromMap(serializedData, provider)
+        : HubEventElementWithMetadata.fromMap(serializedData, provider);
     modelName = serializedData["modelName"] as String;
   }
 }