+++ packages/amplify_datastore_plugin_interface/lib/src/publicTypes.dart
@@ -4,3 +4,6 @@ export 'types/temporal/temporal_datetime.dart';
 export 'types/temporal/temporal_timestamp.dart';
 
 export 'types/exception/DataStoreException.dart';
+export 'types/exception/DataStoreExceptionMessages.dart';
+
+export 'types/query/query_field.dart';