+++ packages/amplify_storage_plugin_interface/lib/src/List/ListRequest.dart
@@ -16,19 +16,17 @@
 import './ListOptions.dart';
 
 class ListRequest {
-  String path;
-  ListOptions options;
+  String? path;
+  ListOptions? options;
 
   ListRequest({this.path, this.options});
 
   Map<String, dynamic> serializeAsMap() {
-    final Map<String, dynamic> result = <String, dynamic>{};
-    if (path != null) {
-      result['path'] = path;
-    }
-    if (options != null) {
-      result['options'] = options.serializeAsMap();
-    }
+    final Map<String, dynamic> result = {
+      'path': path,
+      'options': options?.serializeAsMap()
+    };
+    result.removeWhere((_, v) => v == null);
     return result;
   }
 }