+++ packages/amplify_storage_plugin_interface/lib/src/Remove/RemoveRequest.dart
@@ -13,21 +13,20 @@
  * permissions and limitations under the License.
  */
 
-import 'package:flutter/foundation.dart';
 import './RemoveOptions.dart';
 
 class RemoveRequest {
   String key;
-  RemoveOptions options;
+  RemoveOptions? options;
 
-  RemoveRequest({@required this.key, this.options});
+  RemoveRequest({required this.key, this.options});
 
   Map<String, dynamic> serializeAsMap() {
-    final Map<String, dynamic> result = <String, dynamic>{};
-    result['key'] = key;
-    if (options != null) {
-      result['options'] = options.serializeAsMap();
-    }
+    final Map<String, dynamic> result = {
+      'key': key,
+      'options': options?.serializeAsMap()
+    };
+    result.removeWhere((_, v) => v == null);
     return result;
   }
 }